Allow listening on port 0
And require listen to be called explicitly.

Note that this is a backward incompatible change.  addPeer depends on
the channel to know its own name, and the name is now only knowable
after the channel is listening.  I have removed support for
options.listening and server.listening is now a flag indicating that the
server is listening.  Tests have been adjusted to listen, and those that
required a random, available port, now use port 0 and need not wonder
whether the port was actually available.