Include full Thrift payload in Thrift calls
Currently, we skip out on the method name, message type, and sequence ID to avoid duplicate information in the payload. That avoidance of duplication has the result of complicating the Go and Python implementations a lot because they rely on the code generated by the Apache Thrift compiler.

By including this information in the payload like a standard Thrift request, we get to simplify the client implementations significantly without affecting end users in any way.