Add new streaming generator.
A new interface is generated which embeds the standard interfaces
for both the client and the server.

Clients only need to create a single streaming client and can also
use the non-streaming methods.
Servers are more complicated right now - the non-streaming part
and the streaming part must be registered separately. However, a
single handler can be used for both parts of the server.

This code does not support:
 - non struct return types
 - Thrift exceptions