POC - Add support to register "other" channels from a top level channel.
This will enable us to solve scenarios where we need 2 or more
channels (for services) listening on the same host:port but with their own
handlers.
For example:
* Register a new tchannel, say serviceACh, for "serviceA" listening
  on "host:port" with its own handlers
* Now register another tchannel, named serviceBCh which hangs off of
  the above tchannel (serviceACh) but with its own handlers

This patch simply creates the "other" channel and makes sure the top
level channel is used for all the connection maintanence logic.

The patch also modifies the "ping" example to use this "Other"
channel.

TODO:
* More targetted tests
* Cleanup