[tchannel-go] - Allow subchannel handler registrations
This patch allows the subchannels to register their own set of handlers
independent of the topchannel. They will still be using the topchannel's
host/port for all the traffic.

The connection management involves maintaining a pointer to the
subchannel map (similar to handler map). When a request comes in, we
first look in the handler map to find the handler, if it fails, we try
the subchannel map to see if we can get one.
(In case of a non-existant handler, we will effectively be doing 3 O(1)
lookups. If all the traffic comes through subchannel servicename/op map we will
have one unnecessary lookup on the topchannel)

TODO:
----
* Add peer support for subchannels - right now they are using the same
  peer list as the topchannel